Re-arrange layout
This puts the repository as the last column so that it can be any
size.

I'm suggesting this because I'm using images from
docker-registry.wikimedia.org and the default layout hides the actual
name of the docker image.  Placing it in the last column allows Emacs
to display the whole name no matter how long it is.